+++ NumeralSystems.Net/NumeralSystems.Net/Utils/Convert/Float.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace NumeralSystems.Net.Utils
+{
+    public static partial class Convert
+    {
+        public static float ToFloat(this bool[] s)
+        {
+            var bytes = ToByteArray(s);
+            return BitConverter.ToSingle(bytes);
+        }
+    }
+}
